Implement a card game hook
